requirements.txt
fastapi
uvicorn
pytest
httpx
pytest-asyncio
PyJWT
pwdlib[argon2]
python-dotenv
pydantic[email]
asyncpg
sqlalchemy
greenlet
alembic